apps/backend/pyproject.toml
[project]
name = "janus-agent"
version = "0.1.0"
description = "Janus AI Agent Backend"
requires-python = ">=3.10"
dependencies = [
    "flask==3.0.2",
    "flask-cors==4.0.0",
    "flask-sqlalchemy==3.1.1",
    "psycopg2-binary==2.9.9",
    "langchain==0.1.13",
    "langchain-openai==0.1.1",
    "langchain-pinecone==0.0.3",
    "langgraph==0.0.30",
    "pinecone>=5.0.0,<8.0.0",
    "llama-index-core>=0.12.0,<0.15",
    "llama-index-vector-stores-pinecone>=0.3.0",
    "llama-index-embeddings-openai>=0.2.0",
    "llama-index-readers-file>=0.2.0",
    "pypdf>=4.0.0",
    "python-dotenv==1.0.1",
    "gunicorn==21.2.0",
    "requests==2.31.0",
    "google-api-python-client==2.122.0",
    "google-auth-httplib2==0.2.0",
    "google-auth-oauthlib==1.2.0",
    "pytz",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-flask>=1.3.0",
    "ruff>=0.3.3",
    "pre-commit>=3.6.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4", "SIM", "ARG"]
ignore = ["ARG001", "ARG002", "B008"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
